made some minor changes and fixed the format of backend seperated all the things and created seperate controller folder routes folder etc
