+++ backend/middlewares/errorHandler.js
@@ -1,0 +1,12 @@
+const errorHandler = (err, req, res, next) => {
+  console.error(err.stack);
+  res.status(500).json({
+    message: "Something went wrong!",
+    error:
+      process.env.NODE_ENV === "development"
+        ? err.message
+        : "Internal server error",
+  });
+};
+
+module.exports = errorHandler;
